reuse refractor
